backendFlask/requirements.txt
# Flask framework
Flask==3.1.2
Flask-Cors==6.0.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1

# SQLAlchemy and database
SQLAlchemy==2.0.46
alembic==1.18.3
pyodbc==5.3.0

# Utilities
python-dotenv==1.2.1

# Dependencies (automatically installed with Flask)
blinker==1.9.0
click==8.3.1
colorama==0.4.6
greenlet==3.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
typing_extensions==4.15.0
Werkzeug==3.1.5
waitress==3.0.0